refactor: extract BatchDeduplicator and move heartbeat pruning to heartbeat-filter
Phase 2 of the engine.ts decomposition (#867). The after-turn batch
replay dedup cluster (frontier alignment + overlap dedup + oversized and
suffix fallbacks) moves to src/batch-dedup.ts (BatchDeduplicator,
constructed with the conversation store and logger). pruneHeartbeatOkTurns
becomes a free function in heartbeat-filter.ts taking the conversation
store. Bodies unchanged; fidelity tests that spied on the private dedup
method now spy on the service.

Full suite: 1279 tests passing.
