+++ test/engine-fidelity.test.ts
@@ -2481,12 +2481,14 @@ describe("LcmContextEngine fidelity and token budget", () => {
       compaction: {
         evaluate: (conversationId: number, tokenBudget: number, observed?: number) => Promise<unknown>;
       };
-      deduplicateAfterTurnBatch: (
-        sessionId: string,
-        sessionKey: string | undefined,
-        messages: AgentMessage[],
-        opts: unknown,
-      ) => Promise<AgentMessage[]>;
+      batchDeduplicator: {
+        deduplicateAfterTurnBatch: (
+          sessionId: string,
+          sessionKey: string | undefined,
+          messages: AgentMessage[],
+          opts: unknown,
+        ) => Promise<AgentMessage[]>;
+      };
       scheduleDeferredCompactionDebtDrain: (params: unknown) => void;
     };
 
@@ -2494,7 +2496,7 @@ describe("LcmContextEngine fidelity and token budget", () => {
       sessionId,
       message: makeMessage({ role: "user", content: "seed message" }),
     });
-    vi.spyOn(privateEngine, "deduplicateAfterTurnBatch").mockResolvedValue([]);
+    vi.spyOn(privateEngine.batchDeduplicator, "deduplicateAfterTurnBatch").mockResolvedValue([]);
     vi.spyOn(privateEngine.compaction, "evaluate").mockResolvedValue({
       shouldCompact: true,
       reason: "threshold",
@@ -2534,19 +2536,21 @@ describe("LcmContextEngine fidelity and token budget", () => {
       compaction: {
         evaluate: (conversationId: number, tokenBudget: number, observed?: number) => Promise<unknown>;
       };
-      deduplicateAfterTurnBatch: (
-        sessionId: string,
-        sessionKey: string | undefined,
-        messages: AgentMessage[],
-        opts: unknown,
-      ) => Promise<AgentMessage[]>;
+      batchDeduplicator: {
+        deduplicateAfterTurnBatch: (
+          sessionId: string,
+          sessionKey: string | undefined,
+          messages: AgentMessage[],
+          opts: unknown,
+        ) => Promise<AgentMessage[]>;
+      };
     };
 
     await engine.ingest({
       sessionId,
       message: makeMessage({ role: "user", content: "seed message" }),
     });
-    vi.spyOn(privateEngine, "deduplicateAfterTurnBatch").mockResolvedValue([]);
+    vi.spyOn(privateEngine.batchDeduplicator, "deduplicateAfterTurnBatch").mockResolvedValue([]);
     vi.spyOn(privateEngine.compaction, "evaluate").mockResolvedValue({
       shouldCompact: false,
       reason: "below threshold",
